Autoloader support introduced, with the \Mesamatrix namespace
Upon including lib/base.php, an autoloader will be registered, so any access
to a \Mesamatrix\* class does not require an explicit include.

Each PHP file now contains one class only, so oglparser.php has been split
into its multiple components and the addToHints function has been refactored
into a separate class, an instance of which is passed between the various
\Mesamatrix\Parser\* classes.
