+++ http/index.php
@@ -22,12 +22,11 @@
 // Common code.
 
 require_once "../lib/base.php";
-require_once "hints.php";
 
 ///////////////////////////////////////
 // File code.
 
-$gl3Path = MesaMatrix::path(MesaMatrix::$config->getValue("info", "xml_file"));
+$gl3Path = Mesamatrix::path(Mesamatrix::$config->getValue("info", "xml_file"));
 
 // Read "xml_file".
 $xml = simplexml_load_file($gl3Path);
@@ -65,7 +64,7 @@ usort($glVersions, function($a, $b)
         }
     });
 
-$hints = new Hints();
+$hints = new Mesamatrix\Hints();
 foreach($glVersions as $glVersion)
 {
     foreach($glVersion->extension as $ext)
@@ -118,12 +117,12 @@ foreach($xml->drivers->vendor as $vendor)
 <html>
     <head>
         <meta charset="utf-8"/>
-        <meta name="description" content="<?= MesaMatrix::$config->getValue("info", "description") ?>"/>
+        <meta name="description" content="<?= Mesamatrix::$config->getValue("info", "description") ?>"/>
 
-        <title><?= MesaMatrix::$config->getValue("info", "title") ?></title>
+        <title><?= Mesamatrix::$config->getValue("info", "title") ?></title>
 
         <link rel="shortcut icon" href="images/gears.png" />
-        <link href="css/style.css?v=<?= MesaMatrix::$config->getValue("info", "version") ?>" rel="stylesheet" type="text/css" media="all"/>
+        <link href="css/style.css?v=<?= Mesamatrix::$config->getValue("info", "version") ?>" rel="stylesheet" type="text/css" media="all"/>
         <link href="css/tipsy.css" rel="stylesheet" type="text/css" media="all" />
         <script src="js/jquery-1.11.1.min.js"></script>
         <script src="js/jquery.tipsy.js"></script>
@@ -131,7 +130,7 @@ foreach($xml->drivers->vendor as $vendor)
     </head>
     <body>
         <h1>Last commits</h1>
-        <p><b>Last git update:</b> <script>document.write(getLocalDate("<?= $lastGitUpdate ?>"));</script><noscript><?= $lastGitUpdate ?></noscript> (<a href="<?= MesaMatrix::$config->getValue("git", "web")."/log/docs/GL3.txt" ?>">see the log</a>)</p>
+        <p><b>Last git update:</b> <script>document.write(getLocalDate("<?= $lastGitUpdate ?>"));</script><noscript><?= $lastGitUpdate ?></noscript> (<a href="<?= Mesamatrix::$config->getValue("git", "web")."/log/docs/GL3.txt" ?>">see the log</a>)</p>
 <?php
 foreach($xml->commits->commit as $commit) {
     $commitDate = date(DATE_RFC2822, (int) $commit["timestamp"]);
@@ -141,7 +140,7 @@ foreach($xml->commits->commit as $commit) {
             <noscript><?= $commitDate ?></noscript>:
         </div>
         <div class="commitText">
-            <a href="<?= MesaMatrix::$config->getValue("git", "web")."/commit/".MesaMatrix::$config->getValue("git", "gl3")
+            <a href="<?= Mesamatrix::$config->getValue("git", "web")."/commit/".Mesamatrix::$config->getValue("git", "gl3")
                 ?>?id=<?= $commit["hash"] ?>"><?= $commit["subject"] ?></a>
         </div>
 <?php
@@ -307,11 +306,11 @@ for($i = 0; $i < $numHints; $i++)
         <h1>How to help</h1>
         <p>If you find this page useful and want to help, you can report issues, or <a href="https://github.com/MightyCreak/mesamatrix">grab the code</a> and add whatever feature you want.</p>
 <?php
-if(MesaMatrix::$config->getValue("flattr", "enabled"))
+if(Mesamatrix::$config->getValue("flattr", "enabled"))
 {
 ?>
         <p>You can click here too, if you want to Flattr me:</p>
-        <p><script id='fb5dona'>(function(i){var f,s=document.getElementById(i);f=document.createElement('iframe');f.src='//api.flattr.com/button/view/?uid=<?= MesaMatrix::$config->getValue("flattr", "id") ?>&url='+encodeURIComponent(document.URL)+'&title='+encodeURIComponent('<?= MesaMatrix::$config->getValue("info", "title") ?>')+'&description='+encodeURIComponent('<?= MesaMatrix::$config->getValue("info", "description") ?>')+'&language='+encodeURIComponent('<?= MesaMatrix::$config->getValue("flattr", "language") ?>')+'&tags=<?= MesaMatrix::$config->getValue("flattr", "tags") ?>';f.title='Flattr';f.height=62;f.width=55;f.style.borderWidth=0;s.parentNode.insertBefore(f,s);})('fb5dona');</script></p>
+        <p><script id='fb5dona'>(function(i){var f,s=document.getElementById(i);f=document.createElement('iframe');f.src='//api.flattr.com/button/view/?uid=<?= Mesamatrix::$config->getValue("flattr", "id") ?>&url='+encodeURIComponent(document.URL)+'&title='+encodeURIComponent('<?= Mesamatrix::$config->getValue("info", "title") ?>')+'&description='+encodeURIComponent('<?= Mesamatrix::$config->getValue("info", "description") ?>')+'&language='+encodeURIComponent('<?= Mesamatrix::$config->getValue("flattr", "language") ?>')+'&tags=<?= Mesamatrix::$config->getValue("flattr", "tags") ?>';f.title='Flattr';f.height=62;f.width=55;f.style.borderWidth=0;s.parentNode.insertBefore(f,s);})('fb5dona');</script></p>
 <?php
 }
 ?>
@@ -324,13 +323,13 @@ if(MesaMatrix::$config->getValue("flattr", "enabled"))
             <li>Wikipedia (en): <a href="https://en.wikipedia.org/wiki/Nouveau_%28software%29" title="Nouveau (software)">Nouveau (software)</a></li>
         </ul>
         <h1>Sources</h1>
-        <p><b>This page is generated from:</b> <a href="<?= MesaMatrix::$config->getValue("git", "web")."/tree/".MesaMatrix::$config->getValue("git", "gl3") ?>"><?= MesaMatrix::$config->getValue("git", "web")."/tree/".MesaMatrix::$config->getValue("git", "gl3") ?></a></p>
+        <p><b>This page is generated from:</b> <a href="<?= Mesamatrix::$config->getValue("git", "web")."/tree/".Mesamatrix::$config->getValue("git", "gl3") ?>"><?= Mesamatrix::$config->getValue("git", "web")."/tree/".Mesamatrix::$config->getValue("git", "gl3") ?></a></p>
         <p>If you want to report a bug or simply to participate in the project, feel free to get the sources:
-        <a href="<?= MesaMatrix::$config->getValue("info", "project_url") ?>"><?= MesaMatrix::$config->getValue("info", "project_url") ?></a></p>
+        <a href="<?= Mesamatrix::$config->getValue("info", "project_url") ?>"><?= Mesamatrix::$config->getValue("info", "project_url") ?></a></p>
         <p><a href="http://www.gnu.org/licenses/"><img src="https://www.gnu.org/graphics/gplv3-127x51.png" alt="Logo GPLv3" /></a></p>
         <h1>Authors</h1>
         <ul>
-<?php foreach (MesaMatrix::$config->getValue("info", "authors") as $k => $v) {
+<?php foreach (Mesamatrix::$config->getValue("info", "authors") as $k => $v) {
     if (is_string($k)) { ?>
             <li><a href="<?= $v ?>"><?= $k ?></a></li>
 <?php } else { ?>
