+++ lib/parser/oglmatrix.php
@@ -1,0 +1,41 @@
+<?php
+/*
+ * Copyright (C) 2014 Romain "Creak" Failliot.
+ *
+ * This file is part of mesamatrix.
+ *
+ * mesamatrix is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * any later version.
+ *
+ * mesamatrix is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with mesamatrix. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+namespace Mesamatrix\Parser;
+
+class OglMatrix
+{
+    public function __construct()
+    {
+        $this->glVersions = array();
+    }
+
+    public function addGlVersion($glVersion)
+    {
+        array_push($this->glVersions, $glVersion);
+    }
+
+    public function getGlVersions()
+    {
+        return $this->glVersions;
+    }
+
+    private $glVersions;
+};
